+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.706] - 2024-05-26 - progressing the restoration of the best practice analysis feature and prtg sensor feature
+
+## [0.0.705] - 2024-05-26 - progressing the best practice analysis feature restoration
+
 ## [0.0.704] - 2024-05-25 - bugfix accidentally commented -and
 
 ## [0.0.703] - 2024-05-25 - bugfix incorrect use of $_ should be $ACE